add deadline in project

+++ ViewComponents/Project/Project.cs
@@ -14,7 +14,8 @@ namespace NoteMakingApp.ViewComponents.Project
 {
     public partial class Project : UserControl
     {
-
+        int _year = 0, _day = 0, _hour = 0, _minute = 0, _second = 0;
+        string hour, minute, second;
         List<ItemProjects> item;
         ItemProjects save;
         List<ItemProjects> newitem;
@@ -31,6 +32,7 @@ namespace NoteMakingApp.ViewComponents.Project
             save = new ItemProjects();
             deletedItem = new List<ItemProjects>();
             InitializeComponent();
+            this.lbTime.Visible = false;
         }
 
 
@@ -92,6 +94,7 @@ namespace NoteMakingApp.ViewComponents.Project
             this.lbProjectName.Visible = false;
             this.txtTittle.Visible = true;
             this.lbComplete.Visible = false;
+            this.lbTime.Visible = false;
             btnOK.Visible = true;
             btnDel.Visible = true;
             btnCancel.Visible = true;
@@ -167,6 +170,7 @@ namespace NoteMakingApp.ViewComponents.Project
             this.txtTittle.Visible = false;
             this.lbComplete.Visible = true;
             this.btnAdd.Visible = false;
+            this.lbTime.Visible = false;
             btnOK.Visible = false;
             btnDel.Visible = false;
             btnCancel.Visible = false;
@@ -184,45 +188,51 @@ namespace NoteMakingApp.ViewComponents.Project
             if (txtTittle.Text == "")
                 MessageBox.Show("Không thể để trống!", "Nhập lại", MessageBoxButtons.OK, MessageBoxIcon.Error);
             else
-            {
-                DataHandle.getInstance().EditProject(txtTittle.Text, item);
-                foreach (ItemProjects i in newitem)
-                    DataHandle.getInstance().CreateNewProject(txtTittle.Text, i.Content, i.STT, Convert.ToInt32(i.check));
-                foreach (ItemProjects i in deletedItem)
-                    DataHandle.getInstance().DeleteItemInProject(txtTittle.Text, i.STT);
-
-                double x = 0;
-                double y = 0;
-                string z;
-                Form1.type = 4;
-
-                this.Dispose();
-
-                Projects p = DataHandle.getInstance().GetDataProject();
-
-                foreach (ItemProjects i in p.item)
-                {
-                    if (i.check == true)
-                        x++;
-                    y++;
-                }
-                x = ((x / y) * 100);
-                if (x.ToString().Length > 5)
-                {
-                    z = x.ToString().Substring(0, 5);
-
-                }
+                if (DataHandle.getInstance().checkEditTittleProject(txtTittle.Text) == false)
+                    MessageBox.Show("Tittle đã tồn tại!", "Nhập lại", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 else
                 {
-                    z = x.ToString();
+                    if (item.Count() == 0 && newitem.Count() == 0)
+                    {
+                        MessageBox.Show("Thêm list to do", "Nhập lại", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                    else
+                    {
+                        DataHandle.getInstance().EditProject(txtTittle.Text, item);
+                        foreach (ItemProjects i in newitem)
+                            DataHandle.getInstance().CreateNewProject(txtTittle.Text, i.Content, i.STT, Convert.ToInt32(i.check));
+                        foreach (ItemProjects i in deletedItem)
+                            DataHandle.getInstance().DeleteItemInProject(txtTittle.Text, i.STT);
 
+                        double x = 0;
+                        double y = 0;
+                        string z;
+                        Form1.type = 4;
+
+                        this.Dispose();
+
+                        Projects p = DataHandle.getInstance().GetDataProject();
+
+                        foreach (ItemProjects i in p.item)
+                        {
+                            if (i.check == true)
+                                x++;
+                            y++;
+                        }
+                        x = ((x / y) * 100);
+                        if (x.ToString().Length > 5)
+                        {
+                            z = x.ToString().Substring(0, 5);
+
+                        }
+                        else
+                        {
+                            z = x.ToString();
+
+                        }
+                        Form1.getInstance().ShowProject(p.Tittle, p.item, z,p.deadline);
+                    }
                 }
-
-
-                Form1.getInstance().ShowProject(p.Tittle, p.item, z);
-
-
-            }
         }
 
         private void btnAdd_Click(object sender, EventArgs g)
@@ -287,6 +297,8 @@ namespace NoteMakingApp.ViewComponents.Project
                 }
             };
         }
+
+        
 
         private void addNewEditItem()
         {
@@ -421,5 +433,83 @@ namespace NoteMakingApp.ViewComponents.Project
             else
                 addNewEditItem();
         }
+
+        private void lbTime_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        public void setDeadline(string time)
+        {
+            DateTime _time = Convert.ToDateTime(time);
+
+            _second = _time.Second - DateTime.Now.Second;
+            _minute = _time.Minute - DateTime.Now.Minute;
+            _hour = _time.Hour - DateTime.Now.Hour;
+            _year = _time.Year - DateTime.Now.Year;
+
+            if (_year == 0)
+                _day = _time.DayOfYear - DateTime.Now.DayOfYear;
+            else
+                _day = _time.DayOfYear - DateTime.Now.DayOfYear + 366;
+
+            checkdatetime();
+            converttostring();
+            lbTime.Visible = true;
+            if (_day < 0)
+            {
+                Timer.Enabled = false;
+                lbTime.Text = "Time-expired!";  
+            }
+            else
+            {
+                lbTime.Text = _day + ":" + hour + ":" + minute + ":" + second;  
+            }
+            lbTime.Location = new System.Drawing.Point(this.Width - this.lbTime.Width - 5, 277);
+            Console.WriteLine(lbTime.Text);
+        }
+
+        public void checkdatetime()
+        {
+            if (_second < 0)
+            {
+                _minute--;
+                _second += 60;
+            }
+
+            if (_minute < 0)
+            {
+                _hour--;
+                _minute += 60;
+            }
+
+            if (_hour < 0)
+            {
+                _day--;
+                _hour += 24;
+            }
+        }
+
+        public void converttostring()
+        {
+            if (_hour < 10)
+                hour = "0" + _hour.ToString();
+            else
+                hour = _hour.ToString();
+
+            if (_minute < 10)
+                minute = "0" + _minute.ToString();
+            else
+                minute = _minute.ToString();
+
+            if (_second < 10)
+                second = "0" + _second.ToString();
+            else
+                second = _second.ToString();
+        }
+
+        
+            
+        
     }
 }
